requirements.txt
streamlit==1.31.0
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
pandas==2.2.2
playwright==1.40.0
googlemaps==4.10.0
